renamed `connectCard` to `connect_card`
