+++ components/connect_card/ConnectCard.tsx
